feat: add iPad layout
